requirements_qt.txt
absl-py==1.3.0
astunparse==1.6.3
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
ConfigSpace==0.6.1
contourpy==1.0.5
crc32c==2.3
cryptography==37.0.1
cycler==0.11.0
dill==0.3.6
dpcpp-cpp-rt==2022.2.0
etils==0.9.0
filelock==3.9.0
flatbuffers==22.10.26
fonttools==4.37.4
gast==0.4.0
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
intel-cmplr-lib-rt==2022.2.0
intel-cmplr-lic-rt==2022.2.0
intel-opencl-rt==2022.2.0
intel-openmp==2022.2.0
joblib==1.2.0
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
mkl
mkl-random
mkl-service
more-itertools==9.0.0
numpy
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.3
Pillow==9.2.0
pip==22.3.1
promise==2.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydevd-pycharm==222.4345.23
pyemd==0.5.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy
seaborn==0.12.2
setuptools==65.6.0
six==1.16.0
tbb==2021.7.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-datasets==4.7.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.11.0
termcolor==2.1.0
tfrecord==1.14.1
threadpoolctl==3.1.0
toml==0.10.2
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
zipp==3.10.0
